add stats for constsum recipients
