+++ src/web/app/components/question-types/question-statistics/constsum-recipients-question-statistics.component.ts
@@ -1,13 +1,10 @@
 import { Component, OnChanges, OnInit } from '@angular/core';
-import {
-  FeedbackConstantSumQuestionDetails,
-  FeedbackConstantSumResponseDetails,
-  FeedbackParticipantType,
-} from '../../../../types/api-output';
 import { DEFAULT_CONSTSUM_OPTIONS_QUESTION_DETAILS } from '../../../../types/default-question-structs';
 import { SortBy } from '../../../../types/sort-properties';
 import { ColumnData, SortableTableCellData } from '../../sortable-table/sortable-table.component';
-import { QuestionStatistics } from './question-statistics';
+import {
+  ConstsumRecipientsQuestionStatisticsCalculation,
+} from './question-statistics-calculation/constsum-recipients-question-statistics-calculation';
 
 /**
  * Statistics for constsum recipients questions.
@@ -17,15 +14,8 @@ import { QuestionStatistics } from './question-statistics';
   templateUrl: './constsum-recipients-question-statistics.component.html',
   styleUrls: ['./constsum-recipients-question-statistics.component.scss'],
 })
-export class ConstsumRecipientsQuestionStatisticsComponent
-    extends QuestionStatistics<FeedbackConstantSumQuestionDetails, FeedbackConstantSumResponseDetails>
+export class ConstsumRecipientsQuestionStatisticsComponent extends ConstsumRecipientsQuestionStatisticsCalculation
     implements OnInit, OnChanges {
-
-  emailToTeamName: Record<string, string> = {};
-  emailToName: Record<string, string> = {};
-  pointsPerOption: Record<string, number[]> = {};
-  totalPointsPerOption: Record<string, number> = {};
-  averagePointsPerOption: Record<string, number> = {};
 
   columnsData: ColumnData[] = [];
   rowsData: SortableTableCellData[][] = [];
@@ -42,38 +32,6 @@ export class ConstsumRecipientsQuestionStatisticsComponent
   ngOnChanges(): void {
     this.calculateStatistics();
     this.getTableData();
-  }
-
-  private calculateStatistics(): void {
-    this.emailToTeamName = {};
-    this.emailToName = {};
-    this.pointsPerOption = {};
-    this.totalPointsPerOption = {};
-    this.averagePointsPerOption = {};
-
-    const isRecipientTeam: boolean = this.recipientType === FeedbackParticipantType.TEAMS
-        || this.recipientType === FeedbackParticipantType.TEAMS_EXCLUDING_SELF;
-
-    for (const response of this.responses) {
-      const identifier: string = isRecipientTeam ? response.recipient : (response.recipientEmail || response.recipient);
-
-      this.pointsPerOption[identifier] = this.pointsPerOption[identifier] || [];
-      this.pointsPerOption[identifier].push(response.responseDetails.answers[0]);
-
-      if (!this.emailToTeamName[identifier]) {
-        this.emailToTeamName[identifier] = isRecipientTeam ? '' : response.recipientTeam;
-      }
-      if (!this.emailToName[identifier]) {
-        this.emailToName[identifier] = response.recipient;
-      }
-    }
-
-    for (const recipient of Object.keys(this.pointsPerOption)) {
-      const answers: number[] = this.pointsPerOption[recipient];
-      const sum: number = answers.reduce((a: number, b: number) => a + b, 0);
-      this.totalPointsPerOption[recipient] = sum;
-      this.averagePointsPerOption[recipient] = +(answers.length === 0 ? 0 : sum / answers.length).toFixed(2);
-    }
   }
 
   private getTableData(): void {
